brute force approach of finding closest point

+++ core/math/geometry/Geometry.cpp
@@ -1,8 +1,8 @@
 #include "Geometry.h"
 
 PointPair Geometry::calculateCenter(PointPair p, DimensionsPair dim) {
-	float wMid = dim.first / 2;
-	float hMid = dim.second / 2;
+	int32_t wMid = dim.first / 2;
+	int32_t hMid = dim.second / 2;
 
 	return std::make_pair(p.first + wMid, p.second + hMid);
 }
@@ -14,9 +14,9 @@ bool Geometry::isInRect(PointPair origin, DimensionsPair dim, PointPair query) {
 			&& query.second <= origin.second + dim.second;
 }
 
-PointPair Geometry::lerp(PointPair a, PointPair b, float step) {
-	float x = a.first + step;
-	float y = (a.second * (b.first - x) + b.second * (x - a.first)) / (b.first - a.first);
+PointPair Geometry::lerp(PointPair a, PointPair b, int32_t step) {
+	int32_t x = a.first + step;
+	int32_t y = (a.second * (b.first - x) + b.second * (x - a.first)) / (b.first - a.first);
 
 	return std::make_pair(x, y);
 }
